GiniBankSDK+Pinning: Fix XCFrameworks zip
The zip needs to be created on MacOS by selecting all the XCFrameworks and compressing them without
putting them into a folder.

If we put the XCFrameworks into a folder and then zip that folder then the `unzip` command Cocoapods
uses will put all the XCFrameworks into a subfolder with that name. This breaks adding the
XCFrameworks to the project because Cocoapods can't find XCFrameworks in subfolders.